Backend/requirements.txt
Flask>=2.0 # Specify Flask version, adjust if needed
Flask-CORS>=3.0 # Add Flask-CORS for handling Cross-Origin requests
psycopg2-binary
bcrypt
Flask-JWT-Extended>=4.0 # For JWT token authentication
